+++ processor/src/test/fixtures/expected/com/example/prop/React4j_MultiContextPropModel.java
@@ -1,0 +1,130 @@
+package com.example.prop;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.annotations.JsConstructor;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.React;
+import react4j.ReactNode;
+import react4j.internal.ComponentConstructorFunction;
+import react4j.internal.NativeComponent;
+import react4j.internal.OnComponentWillUnmount;
+import react4j.internal.OnShouldComponentUpdate;
+
+@SuppressWarnings("Arez:UnnecessaryAllowEmpty")
+@ArezComponent(
+    name = "MultiContextPropModel",
+    disposeNotifier = Feature.DISABLE,
+    allowEmpty = true,
+    dagger = Feature.DISABLE
+)
+@Generated("react4j.processor.React4jProcessor")
+abstract class React4j_MultiContextPropModel extends MultiContextPropModel {
+  React4j_MultiContextPropModel(@Nonnull final NativeComponent $$react4j$$_nativeComponent) {
+    bindComponent( $$react4j$$_nativeComponent );
+  }
+
+  @Nonnull
+  private static ComponentConstructorFunction getConstructorFunction() {
+    final ComponentConstructorFunction componentConstructor = ( React.shouldStoreDebugDataAsState() || React.shouldValidatePropValues() ) ? NativeReactComponent::new : LiteNativeReactComponent::new;
+    if ( React.enableComponentNames() ) {
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "MultiContextPropModel" );
+    }
+    return componentConstructor;
+  }
+
+  @Override
+  String getStringContextValue() {
+    if ( React.shouldCheckInvariants() ) {
+      return null != props().getAsAny( Props.stringContextValue ) ? props().getAsAny( Props.stringContextValue ).asString() : null;
+    } else {
+      return Js.uncheckedCast( props().getAsAny( Props.stringContextValue ) );
+    }
+  }
+
+  @Override
+  int getIntContextValue() {
+    return props().getAsAny( Props.intContextValue ).asInt();
+  }
+
+  private boolean $$react4j$$_shouldComponentUpdate(
+      @Nullable final JsPropertyMap<Object> nextProps) {
+    assert null != nextProps;
+    final JsPropertyMap<Object> props = props();
+    if ( !Js.isTripleEqual( props.get( Props.stringContextValue ), nextProps.get( Props.stringContextValue ) ) ) {
+      return true;
+    }
+    if ( !Js.isTripleEqual( props.get( Props.intContextValue ), nextProps.get( Props.intContextValue ) ) ) {
+      return true;
+    }
+    return false;
+  }
+
+  private void $$react4j$$_componentWillUnmount() {
+    ((Arez_React4j_MultiContextPropModel) this).dispose();
+  }
+
+  static final class Factory {
+    @Nonnull
+    static final ComponentConstructorFunction TYPE = getConstructorFunction();
+  }
+
+  static final class Props {
+    static final String stringContextValue = React.shouldMinimizePropKeys() ? "a" : "stringContextValue";
+
+    static final String intContextValue = React.shouldMinimizePropKeys() ? "b" : "intContextValue";
+  }
+
+  private static final class LiteNativeReactComponent extends NativeComponent implements OnShouldComponentUpdate {
+    @Nonnull
+    private final React4j_MultiContextPropModel $$react4j$$_component;
+
+    @JsConstructor
+    LiteNativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
+      super( props );
+      $$react4j$$_component = new Arez_React4j_MultiContextPropModel( this );
+    }
+
+    @Override
+    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextProps) {
+      return $$react4j$$_component.$$react4j$$_shouldComponentUpdate( nextProps );
+    }
+
+    @Override
+    @Nullable
+    public final ReactNode render() {
+      return $$react4j$$_component.render();
+    }
+  }
+
+  private static final class NativeReactComponent extends NativeComponent implements OnShouldComponentUpdate, OnComponentWillUnmount {
+    @Nonnull
+    private final React4j_MultiContextPropModel $$react4j$$_component;
+
+    @JsConstructor
+    NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
+      super( props );
+      $$react4j$$_component = new Arez_React4j_MultiContextPropModel( this );
+    }
+
+    @Override
+    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextProps) {
+      return $$react4j$$_component.$$react4j$$_shouldComponentUpdate( nextProps );
+    }
+
+    @Override
+    public final void componentWillUnmount() {
+      $$react4j$$_component.$$react4j$$_componentWillUnmount();
+    }
+
+    @Override
+    @Nullable
+    public final ReactNode render() {
+      return $$react4j$$_component.render();
+    }
+  }
+}